factor out HybridisationPC and let it take any form
