+++ case_studies/shallow_water/blockpc/lswe_rblock_scpc.py
@@ -5,113 +5,12 @@ import utils.shallow_water.gravity_bumps as lcase
 from utils import shallow_water as swe
 from utils.planets import earth
 from utils import units
+from utils.hybridisation import HybridisedSCPC  # noqa: F401
 
 import numpy as np
 
 PETSc.Sys.popErrorHandler()
 Print = PETSc.Sys.Print
-
-
-class HybridisedSCPC(fd.PCBase):
-    def initialize(self, pc):
-        if pc.getType() != "python":
-            raise ValueError("Expecting PC type python")
-
-        from ufl import replace
-        from utils.broken_projections import BrokenHDivProjector
-
-        appctx = self.get_appctx(pc)
-        _, P = pc.getOperators()
-        ctx = P.getPythonContext()
-        test, trial = ctx.a.arguments()
-        tests, trials = fd.split(test), fd.split(trial)
-
-        V = test.function_space()
-        mesh = V.mesh()
-
-        Vu, Vh = V.subfunctions
-        Vub = fd.FunctionSpace(mesh, fd.BrokenElement(Vu.ufl_element()))
-        Tr = fd.FunctionSpace(mesh, "HDivT", Vu.ufl_element().degree())
-        Vtr = Vub*Vh*Tr
-
-        # index of hdiv space
-        iu = 0
-
-        self.projector = BrokenHDivProjector(Vu)
-
-        self.x = fd.Cofunction(V.dual())
-        self.y = fd.Function(V)
-
-        self.xu, self.xh = self.x.subfunctions
-        self.yu, self.yh = self.y.subfunctions
-
-        self.xtr = fd.Cofunction(Vtr.dual())
-        self.ytr = fd.Function(Vtr)
-
-        self.xbu, self.xbh, _ = self.xtr.subfunctions
-        self.ybu, self.ybh, _ = self.ytr.subfunctions
-
-        utrs = fd.TrialFunctions(Vtr)
-        vtrs = fd.TestFunctions(Vtr)
-
-        # break the original form
-        news = (*vtrs[:-1], *utrs[:-1])
-        olds = (*tests, *trials)
-        arg_map = {old: new for old, new in zip(olds, news)}
-        A = replace(ctx.a, arg_map)
-
-        # add the trace bit
-        n = fd.FacetNormal(mesh)
-        A += (
-            fd.jump(vtrs[iu], n)*utrs[-1]('+')
-            + fd.jump(utrs[iu], n)*vtrs[-1]('+')
-        )*fd.dS
-
-        Ltr = self.xtr
-
-        scpc_params = {
-            "mat_type": "matfree",
-            "ksp_type": "preonly",
-            "pc_type": "python",
-            "pc_python_type": "firedrake.SCPC",
-            "pc_sc_eliminate_fields": "0, 1",
-        }
-
-        problem = fd.LinearVariationalProblem(A, Ltr, self.ytr)
-        self.solver = fd.LinearVariationalSolver(
-            problem, appctx=appctx,
-            solver_parameters=scpc_params,
-            options_prefix=pc.getOptionsPrefix()+"hybridscpc")
-
-    def apply(self, pc, x, y):
-        # copy into unbroken vector
-        with self.x.dat.vec_wo as v:
-            x.copy(v)
-
-        # break velocity
-        self.projector.project(self.xu, self.xbu)
-
-        # depth already broken
-        self.xbh.assign(self.xh)
-
-        self.ytr.assign(0)
-        self.solver.solve()
-
-        # mend velocity
-        self.projector.project(self.ybu, self.yu)
-
-        # depth already mended
-        self.yh.assign(self.ybh)
-
-        # copy out to petsc
-        with self.y.dat.vec_ro as v:
-            v.copy(y)
-
-    def update(self, pc):
-        pass
-
-    def applyTranspose(self, pc, x, y):
-        raise NotImplementedError
 
 
 # get command arguments
@@ -172,17 +71,21 @@ lu_params = {
     'pc_factor_mat_solver_type': 'petsc'
 }
 
-sparams = {
-    'ksp': {
-        'monitor': None,
-        'converged_rate': None,
-    },
+hybrid_params = {
     'mat_type': 'matfree',
     'ksp_type': 'preonly',
     'pc_type': 'python',
     'pc_python_type': f'{__name__}.HybridisedSCPC',
     'hybridscpc_condensed_field': lu_params
 }
+
+sparams = {
+    'ksp': {
+        'monitor': None,
+        'converged_rate': None,
+    },
+}
+sparams.update(hybrid_params)
 
 # trace component should have zero rhs
 np.random.seed(args.seed)
