+++ utils/__init__.py
@@ -8,4 +8,4 @@ import utils.diagnostics  # noqa: F401
 import utils.timing  # noqa: F401
 import utils.serial  # noqa: F401
 import utils.misc  # noqa: F401
-from utils.broken_projections import *  # noqa: F401
+from utils.hybridisation import *  # noqa: F401
